requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
requests
python-dotenv
